Hex actually works
Sorry

+++ transport/lcd/lcd.h
@@ -43,5 +43,5 @@ void display_char(char c);
 void display_string(char *str);
 void display_nstring(char *str, uint8_t start, uint8_t n);
 void display_number(uint16_t number);
-void display_hex(uint16_t hex);
+void display_hex(uint8_t hex);
 void display_binary(uint8_t byte);
